Expand extranonce2 size to 64-bit
It's rather unlikely to overflow a 32-bit extranonce2 (requires over 4 Gh performed on CPU; with a 1 minute time between jobs, that's over 50 Mh/s), but it doesn't hurt much

+++ miner.h
@@ -1250,7 +1250,7 @@ struct pool {
 	char *sockaddr_url; /* stripped url used for sockaddr */
 	char *nonce1;
 	size_t n1_len;
-	uint32_t nonce2;
+	uint64_t nonce2;
 	int nonce2sz;
 #ifdef WORDS_BIGENDIAN
 	int nonce2off;
